Adjusted macro explicitly to create a uint8_t.

+++ Components/Z80/Implementation/Z80Internals.h
@@ -239,7 +239,7 @@ void llz80_setF(struct LLZ80ProcessorState *z80, uint8_t value);
 
 */
 #define llz80_calculateParity(v)	\
-	int parity = ~v;\
+	uint8_t parity = ~v;\
 	parity ^= parity >> 4;\
 	parity ^= parity << 2;\
 	parity ^= parity >> 1;\
